packages: Implement decrypt fself in a fly. - Refactor license.

+++ vita3k/packages/src/license.cpp
@@ -22,34 +22,12 @@
 
 #include <emuenv/state.h>
 
+#include <packages/license.h>
+
 #include <util/bytes.h>
 #include <util/log.h>
 
 #include <zrif2rif.h>
-
-struct SceNpDrmLicense {
-    uint16_t version; // 0x00
-    uint16_t version_flag; // 0x02
-    uint16_t type; // 0x04
-    uint16_t flags; // 0x06
-    uint64_t aid; // 0x08
-    char content_id[0x30]; // 0x10
-    uint8_t key_table[0x10]; // 0x40
-    uint8_t key[0x10]; // 0x50
-    uint64_t start_time; // 0x60
-    uint64_t expiration_time; // 0x68
-    uint8_t ecdsa_signature[0x28]; // 0x70
-    uint64_t flags2; // 0x98
-    uint8_t key2[0x10]; // 0xA0
-    uint8_t unk_B0[0x10]; // 0xB0
-    uint8_t openpsid[0x10]; // 0xC0
-    uint8_t unk_D0[0x10]; // 0xD0
-    uint8_t cmd56_handshake[0x14]; // 0xE0
-    uint32_t unk_F4; // 0xF4
-    uint32_t unk_F8; // 0xF8
-    int32_t sku_flag; // 0xFC
-    uint8_t rsa_signature[0x100]; // 0x100
-};
 
 static bool open_license(const fs::path &license_path, SceNpDrmLicense &license_buf) {
     memset(&license_buf, 0, sizeof(SceNpDrmLicense));
@@ -87,34 +65,43 @@ bool copy_license(EmuEnvState &emuenv, const fs::path &license_path) {
     return false;
 }
 
-int32_t get_license_sku_flag(EmuEnvState &emuenv, const std::string &content_id) {
-    int32_t sku_flag;
-    const auto title_id = content_id.substr(7, 9);
+void get_license(EmuEnvState &emuenv, const std::string &title_id, const std::string &content_id) {
+    // Skip if it's not a retail game or already have a license
+    if (!title_id.starts_with("PCS") || emuenv.license.rif.contains(title_id))
+        return;
+
+    // Get license buffer corresponding to the title id
+    auto &license_buf = emuenv.license.rif[title_id];
+    license_buf = {};
+
+    // Open license file
     const auto license_path{ emuenv.pref_path / "ux0/license" / title_id / fmt::format("{}.rif", content_id) };
-    SceNpDrmLicense license_buf;
-    if (open_license(license_path, license_buf)) {
-        sku_flag = byte_swap(license_buf.sku_flag);
-    } else {
+    if (!open_license(license_path, license_buf)) {
+        if (fs::exists(license_path))
+            fs::remove(license_path);
+
+        LOG_WARN("License file is corrupted or missing at: {}, using default value.", license_path);
         const auto RETAIL_APP_PATH{ emuenv.pref_path / "ux0/app" / title_id / "sce_sys/retail/livearea" };
         if (fs::exists(RETAIL_APP_PATH))
-            sku_flag = 1;
+            license_buf.sku_flag = 1;
         else
-            sku_flag = 0;
-        if (fs::exists(license_path)) {
-            fs::remove(license_path);
-            LOG_WARN("License file is corrupted at: {}, using default value.", license_path);
-        }
-    }
-
-    return sku_flag;
+            license_buf.sku_flag = 0;
+    } else
+        license_buf.sku_flag = byte_swap(license_buf.sku_flag); // Convert to little endian
 }
 
 bool create_license(EmuEnvState &emuenv, const std::string &zRIF) {
     fs::create_directories(emuenv.cache_path);
 
-    // Create temp license file
+    // Create a temp license file
     const auto temp_license_path = emuenv.cache_path / "temp_licence.rif";
     fs::ofstream temp_file(temp_license_path, std::ios::out | std::ios::binary);
+    if (!temp_file.is_open()) {
+        LOG_ERROR("Failed to create temp license file at: {}", temp_license_path);
+        return false;
+    }
+
+    // Convert zRIF to RIF
     zrif2rif(zRIF, temp_file);
     auto res = copy_license(emuenv, temp_license_path);
     fs::remove(temp_license_path);
